+++ gemd/vdi/vrt_cpyfm.cc
@@ -178,7 +178,7 @@ void VDI::vrt_cpyfm(Workstation *ws, ClientMsg *cm)
 		if (src.fd_wdwidth * 2 * src.fd_h != ba.size())
 			WARN("Mismatch in size within MFDB (%d != %d)",
 				 src.fd_wdwidth * 2 * src.fd_h,
-				 ba.size());
+				 (int)ba.size());
 
 		memcpy(src.fd_addr, ba.constData(), ba.size());
 		}
@@ -214,7 +214,7 @@ void VDI::vrt_cpyfm(Workstation *ws, ClientMsg *cm)
 		if (dst.fd_wdwidth * 2 * src.fd_h != ba.size())
 			WARN("Mismatch in size within dst MFDB (%d != %d)",
 				 dst.fd_wdwidth * 2 * dst.fd_h,
-				 ba.size());
+				 (int)ba.size());
 		memcpy(dst.fd_addr, ba.constData(), ba.size());
 		}
